+++ packages/core/src/lib/auth/social-account/repository/mikro-orm-social-account.repository.ts
@@ -1,4 +1,4 @@
 import { MikroOrmBaseEntityRepository } from '../../../core/repository/mikro-orm-base-entity.repository';
 import { SocialAccount } from '../social-account.entity';
 
-export class MicroOrmSocialAccountRepository extends MikroOrmBaseEntityRepository<SocialAccount> {}
+export class MikroOrmSocialAccountRepository extends MikroOrmBaseEntityRepository<SocialAccount> {}
